refactored api-server to live inside react app folder structure to get rid of CORS issue...hopefully!
