--- outputs/web-app/v4/api-server/.env.example
@@ -1,25 +1,0 @@
-# Environment variables for Todo API Server
-
-# Server configuration
-PORT=3001
-NODE_ENV=development
-
-# Authentication
-QUIZ_PASSCODE=demo123
-
-# Database configuration (for future implementation)
-# DATABASE_URL=postgresql://username:password@localhost:5432/todo_workflow
-
-# ElevenLabs configuration
-ELEVEN_LABS_API_KEY=your_ELEVEN_LABS_API_KEY_here
-# ELEVENLABS_AGENT_ID=your_agent_id_here
-
-# Logging level
-LOG_LEVEL=info
-
-# CORS settings
-CORS_ORIGIN=http://localhost:3000
-
-# Supabase configuration
-SUPABASE_URL=your_supabase_url_here
-SUPABASE_KEY=your_supabase_api_key_here
