--- outputs/web-app/v4/api-server/.gitignore
@@ -1,6 +1,0 @@
-node_modules/
-.env
-*.log
-.DS_Store
-dist/
-build/
