--- outputs/web-app/v4/api-server/netlify.toml
@@ -1,8 +1,0 @@
-[functions]
-  external_node_modules = ["express"]
-  node_bundler = "esbuild"
-[[redirects]]
-  force = true
-  from = "/api/*"
-  status = 200
-  to = "/.netlify/functions/api/:splat"
