--- outputs/web-app/v4/api-server/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "todo-api-server",
-  "version": "1.0.0",
-  "description": "Basic Express API server for todo workflow and conversation endpoints",
-  "main": "server.js",
-  "scripts": {
-    "start": "node src/server.js",
-    "dev": "nodemon src/server.js",
-    "build": "echo Fake Building Functions",
-    "test": "node test-signed-url.js",
-    "test:quick": "node test-quick.js",
-    "test:step": "node test-quick.js"
-  },
-  "keywords": [
-    "express",
-    "api",
-    "workflow",
-    "conversation"
-  ],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@elevenlabs/elevenlabs-js": "^2.1.0",
-    "@supabase/supabase-js": "^2.0.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "node-fetch": "^2.7.0",
-    "serverless-http": "^3.2.0"
-  },
-  "devDependencies": {
-    "netlify-lambda": "*",
-    "nodemon": "^3.0.1"
-  }
-}
