--- outputs/web-app/v4/api-server/src/handlers/index.js
@@ -1,12 +1,0 @@
-// Export all handlers for easy importing
-const { healthHandler } = require("./health");
-const { getWorkflowHandler } = require("./workflow");
-const { getConversationHandler } = require("./conversation");
-const { getSignedUrlHandler } = require("./agent");
-
-module.exports = {
-  healthHandler,
-  getWorkflowHandler,
-  getConversationHandler,
-  getSignedUrlHandler,
-};
