+++ outputs/web-app/v4/voice-step-flow/api-server/src/handlers/health.js
@@ -1,12 +1,10 @@
 /**
  * Health check endpoint handler
  */
-const healthHandler = (req, res) => {
+export const healthHandler = (req, res) => {
   res.json({
     status: "healthy",
     timestamp: new Date().toISOString(),
     service: "todo-api-server",
   });
 };
-
-module.exports = { healthHandler };
