+++ outputs/web-app/v4/voice-step-flow/api-server/src/handlers/index.js
@@ -1,0 +1,12 @@
+// Export all handlers for easy importing
+import { healthHandler } from "./health.js";
+import { getWorkflowHandler } from "./workflow.js";
+import { getConversationHandler } from "./conversation.js";
+import { getSignedUrlHandler } from "./agent.js";
+
+export {
+  healthHandler,
+  getWorkflowHandler,
+  getConversationHandler,
+  getSignedUrlHandler,
+};
