+++ outputs/web-app/v4/voice-step-flow/api-server/test-signed-url.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const fetch = require("node-fetch");
+import fetch from "node-fetch";
 
 // Configuration
 const SERVER_URL = "http://localhost:3001";
