+++ outputs/web-app/v4/voice-step-flow/netlify.toml
@@ -9,6 +9,7 @@
 [functions]
   external_node_modules = ["express"]
   node_bundler = "esbuild"
+  
 [[redirects]]
   force = true
   from = "/api/*"
