+++ outputs/web-app/v4/voice-step-flow/package.json
@@ -8,7 +8,13 @@
     "build": "vite build",
     "build:dev": "vite build --mode development",
     "lint": "eslint .",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "api-start": "node api-server/src/server.js",
+    "api-dev": "nodemon api-server/src/server.js",
+    "api-build": "echo Fake Building Functions",
+    "api-test": "node api-server/test-signed-url.js",
+    "api-test:quick": "node api-server/test-quick.js",
+    "api-test:step": "node api-server/test-quick.js"
   },
   "dependencies": {
     "@elevenlabs/elevenlabs-js": "^2.1.0",
@@ -41,13 +47,17 @@
     "@radix-ui/react-toggle": "^1.1.0",
     "@radix-ui/react-toggle-group": "^1.1.0",
     "@radix-ui/react-tooltip": "^1.1.4",
+    "@supabase/supabase-js": "^2.49.8",
     "@tanstack/react-query": "^5.56.2",
     "caniuse-lite": "^1.0.30001718",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
+    "cors": "^2.8.5",
     "date-fns": "^3.6.0",
+    "dotenv": "^16.5.0",
     "embla-carousel-react": "^8.3.0",
+    "express": "^5.1.0",
     "input-otp": "^1.2.4",
     "lucide-react": "^0.462.0",
     "next-themes": "^0.3.0",
@@ -78,6 +88,8 @@
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
     "eslint-plugin-react-refresh": "^0.4.9",
     "globals": "^15.9.0",
+    "netlify-lambda": "^2.0.16",
+    "nodemon": "^3.1.10",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.11",
     "typescript": "^5.5.3",
